Add Argon2 test.

+++ test.c
@@ -7,17 +7,17 @@ int pbkdf2_test_vectors(void);
 
 static void pbkdf_performance(const char *test_hash, size_t key_len)
 {
-	uint64_t iter;
+	uint32_t iter;
 
 	crypt_pbkdf_check("pbkdf2", test_hash, "foo", 3, "bar", 3, key_len, &iter);
 	printf("PBKDF2 %-9s (%-3u bits key): %10llu iterations/second\n",
-	       test_hash, key_len * 8, (long long unsigned)iter);
+	       test_hash, key_len * 8, (long unsigned)iter);
 }
 
 int main (int argc, char *argv[])
 {
 	struct utsname uts;
-	
+
 	if (crypt_backend_init(NULL)) {
 		printf("Cannot initialize crypto backend.");
 		return 1;
